requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
langchain==0.1.0
langchain-community==0.0.10
langchain-google-genai==0.0.6
chromadb==0.4.22
google-generativeai==0.3.2
pypdf2==3.0.1
python-docx==1.1.0
pydantic-settings==2.1.0
python-multipart==0.0.6
sqlalchemy==2.0.25
